# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: HackWerk HTTP API
  version: 0.4.0
  description: Interner Vertrag des serverseitig gerenderten HackWerk-Webdienstes.
servers:
  - url: http://localhost:18533
paths:
...
  /calendar:
    get:
      operationId: getCalendarPage
      description: Responsive Tages-/Wochen-/Agendaansicht; Fahrer sehen alle Termine read-only, Admins zusätzlich Warteliste und Einplanungsdialog.
      security: [{ sessionCookie: [] }]
      responses:
        "200": { description: Serverseitig gerenderte Kalenderseite mit lokal gebündeltem FullCalendar Standard. }
        "403": { description: Kalenderberechtigung fehlt. }
  /api/v1/calendar:
    get:
      operationId: listCalendarEvents
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/CalendarFrom"
        - $ref: "#/components/parameters/CalendarTo"
      responses:
        "200":
          description: Begrenzte FullCalendar-Events ohne Telefon/E-Mail.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/CalendarEvent" }
        "403": { description: Kalenderberechtigung fehlt. }
        "422": { description: Zeitraum ungültig oder länger als 93 Tage. }
  /api/v1/calendar/conflicts:
    get:
      operationId: listCalendarConflicts
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: "#/components/parameters/CalendarFrom"
        - $ref: "#/components/parameters/CalendarTo"
        - { name: driver_id, in: query, schema: { type: array, items: { type: string, format: uuid } } }
        - { name: resource_id, in: query, schema: { type: array, items: { type: string, format: uuid } } }
        - { name: exclude_appointment_id, in: query, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Datensparsame aktive Fahrer-/Ressourcenüberschneidungen. }
        "422": { description: Zeitraum oder ID ungültig. }
  /api/v1/calendar/plan:
    post:
      operationId: planWaitlistProposal
      description: Erzeugt Draft, geprüfte Zuweisungen und Proposal; fixiert nie und versendet nichts.
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/PlanAppointmentForm" }
      responses:
        "201": { description: Proposal gespeichert. }
        "403": { description: Adminberechtigung oder CSRF fehlt. }
        "409": { description: Version oder exklusive Reservierung kollidiert. }
        "422": { description: Zeit, Verfügbarkeit, Transport oder Zuweisung ungültig. }
  /api/v1/appointments/{appointmentID}/move:
    parameters: [{ $ref: "#/components/parameters/AppointmentID" }]
    post:
      operationId: moveAppointment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/AppointmentTimeMutation" }
      responses:
        "200": { description: Zeit und neue Version gespeichert; fixierte Bestätigung wieder pending. }
        "403": { description: Nur Administratoren dürfen verschieben. }
        "409": { description: Stale Version oder physische Überschneidung; Browser muss revertieren. }
        "422": { description: Zeit oder Verfügbarkeit ungültig. }
  /api/v1/appointments/{appointmentID}/assign:
    parameters: [{ $ref: "#/components/parameters/AppointmentID" }]
    post:
      operationId: assignAppointment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionForm"
                - $ref: "#/components/schemas/AppointmentAssignments"
      responses:
        "200": { description: Aktive Fahrer/Ressourcen typgerecht zugewiesen und Version erhöht. }
        "403": { description: Nur Administratoren dürfen zuweisen. }
        "409": { description: Stale Version oder exklusive Reservierungskollision. }
        "422": { description: Inaktive/falsche Zuweisung, Transport oder Verfügbarkeit ungültig. }
  /api/v1/appointments/{appointmentID}/propose:
    parameters: [{ $ref: "#/components/parameters/AppointmentID" }]
    post:
      operationId: proposeAppointment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionForm"
                - type: object
                  properties: { override_reason: { type: string, maxLength: 1000 } }
      responses:
        "200": { description: Draft als blockierendes Proposal aktiviert. }
        "403": { description: Nur Administratoren dürfen vorschlagen. }
        "409": { description: Stale Version oder exklusive Reservierungskollision. }
        "422": { description: Zuweisung, Transport oder Verfügbarkeit ungültig. }
  /api/v1/appointments/{appointmentID}/resize:
    parameters: [{ $ref: "#/components/parameters/AppointmentID" }]
    post:
      operationId: resizeAppointment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/AppointmentTimeMutation" }
      responses:
        "200": { description: Dauer und neue Version gespeichert. }
        "403": { description: Nur Administratoren dürfen die Dauer ändern. }
        "409": { description: Stale Version oder physische Überschneidung. }
        "422": { description: Zeit oder Verfügbarkeit ungültig. }
  /api/v1/appointments/{appointmentID}/fix:
    parameters: [{ $ref: "#/components/parameters/AppointmentID" }]
    post:
      operationId: fixAppointment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema: { $ref: "#/components/schemas/VersionForm" }
      responses:
        "200": { description: Termin, Auftrag, Warteliste, Reservierungen, Audit und appointment.fixed-Outboxevent atomar gespeichert; kein HTTP-Versand. }
        "403": { description: Nur Administratoren dürfen fixieren. }
        "409": { description: Stale Version oder Reservierungskonflikt. }
        "422": { description: Fahrer, Hackressource, Transport oder Verfügbarkeit nicht fixierbar. }
  /api/v1/appointments/{appointmentID}/cancel:
    parameters: [{ $ref: "#/components/parameters/AppointmentID" }]
    post:
      operationId: cancelAppointment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionForm"
                - type: object
                  properties: { reason: { type: string, maxLength: 1000 } }
      responses:
        "200": { description: Termin storniert und Reservierungen atomar freigegeben. }
        "403": { description: Nur Administratoren dürfen absagen. }
        "409": { description: Stale Version. }
        "422": { description: Status oder erforderlicher Grund ungültig. }
  /api/v1/appointments/{appointmentID}/complete:
    parameters: [{ $ref: "#/components/parameters/AppointmentID" }]
    post:
      operationId: completeAppointment
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              allOf:
                - $ref: "#/components/schemas/VersionForm"
                - type: object
                  properties: { override_reason: { type: string, maxLength: 1000 } }
      responses:
        "200": { description: Fixierter gestarteter Termin durch Admin oder zugewiesenen berechtigten Fahrer abgeschlossen. }
        "403": { description: Abschlussrecht oder Objektzuweisung fehlt. }
        "409": { description: Stale Version. }
        "422": { description: Status/Zeitpunkt ungültig oder Admin-Override fehlt. }
...
components:
  parameters:
...
    CalendarFrom:
      name: from
      in: query
      required: true
      schema: { type: string, format: date-time }
    CalendarTo:
      name: to
      in: query
      required: true
      schema: { type: string, format: date-time }
    AppointmentID:
      name: appointmentID
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
    CalendarEvent:
      type: object
      additionalProperties: false
      required: [id, title, start, end, editable, className, color, contrastColor, extendedProps]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
        editable: { type: boolean, description: Nur für Admin und editierbare Lifecycle-Zustände true. }
        className: { type: string, description: FullCalendar-v7-Klassen für Lifecycle und Bestätigung. }
        color: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        contrastColor: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        extendedProps:
          type: object
          description: Auftrag, Ort, m³, Status, Zuweisungen, Version und erlaubte Aktionen; niemals Telefon oder E-Mail.
          additionalProperties: true
    PlanAppointmentForm:
      type: object
      required: [csrf_token, job_id, starts_at, duration_minutes, driver_id, primary_driver_id, chipper_resource_id]
      properties:
        csrf_token: { type: string, writeOnly: true }
        job_id: { type: string, format: uuid }
        starts_at: { type: string, format: local-date-time, description: Strikte Europe/Vienna-Ortszeit aus datetime-local. }
        duration_minutes: { type: integer, minimum: 15, maximum: 10080 }
        driver_id: { type: array, minItems: 1, items: { type: string, format: uuid } }
        primary_driver_id: { type: string, format: uuid }
        chipper_resource_id: { type: array, minItems: 1, items: { type: string, format: uuid } }
        transport_resource_id: { type: array, items: { type: string, format: uuid } }
        trailer_resource_id: { type: array, items: { type: string, format: uuid } }
        override_reason: { type: string, maxLength: 1000 }
    AppointmentTimeMutation:
      type: object
      required: [csrf_token, version, starts_at, ends_at]
      properties:
        csrf_token: { type: string, writeOnly: true }
        version: { type: integer, minimum: 1 }
        starts_at: { type: string, format: date-time }
        ends_at: { type: string, format: date-time }
        override_reason: { type: string, maxLength: 1000 }
    AppointmentAssignments:
      type: object
      required: [driver_id, primary_driver_id, chipper_resource_id]
      properties:
        driver_id: { type: array, minItems: 1, items: { type: string, format: uuid } }
        primary_driver_id: { type: string, format: uuid }
        chipper_resource_id: { type: array, minItems: 1, items: { type: string, format: uuid } }
        transport_resource_id: { type: array, items: { type: string, format: uuid } }
        trailer_resource_id: { type: array, items: { type: string, format: uuid } }
        override_reason: { type: string, maxLength: 1000 }
...
    VersionForm:
      type: object
      required: [csrf_token, version]
      properties:
        csrf_token: { type: string, writeOnly: true }
        version: { type: integer, minimum: 1 }
        customer_id: { type: string, format: uuid }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hackplan_session